Ignore 'ERROR' updates
This seems to happen once in while, but we cannot just stop processing the
content here: there might be more coming that we do need to process.

One instance that I saw looks like this:
'Error while watching: 401: The event in requested index is outdated and cleared (the requested history has been cleared [3943/3836]) [4942], restarting watch'

Note also that 'close' doesn't exist on a Readable, and in fact there is no
method at all to stop listening[1]. Ultimately this is ok in this case, as
we mustn't ignore events anyways.

[1] http://stackoverflow.com/questions/19277094/how-to-close-a-readable-stream-before-end