Inject the queue ARN into policy documents
AWS requires the policy to specify the ARN. This complicates the live of users of the aws-resource-service,
as we're purposefully hiding these details. In the case of updating a policy for a queue we can determine
the ARN, and use it as 'Resource' field of policy statements that do not specify any value there.

For creating new queues this is seemingly handled by AWS automatically - or AWS has a bug (CloudFormation for example
requires to separate the "create" and "set policy" steps)