Restart the watch from the previously seen resource version
The comment 're-syncing everything' applied to a previous version where we would actually
do a LIST request before starting a watch, We're no longer doing this, but that means in turn
we need to remember where to start from.

Note that this breaks a side-effect that was actually "ok": if we failed to create a queue
we would eventually retry when we reprocessed all events again after a watch restart. This needs
to be handled differently by actually having a local "desired" state and keeping track of where
we're still having to do work to get there.