List resources before watching
Looking at the resource version of an update doesn't really make sense for us: the version
of the list is completely unrelated to that. We also cannot make any assumptions at all about
the type of a resource version [1], and need to treat it as a opaque value. Finally, from
discussion on the kubernetes issue tracker it also seems that resourceVersion "0" either shouldn't
work, or just means "whatever you have right now".

Explicitly list the queues known, treat these as initial syncing updates, and keep the resource
version from that list. Then start watching for changes on that list, from the version we now know.

This may lead to no-op updates (it seems that we do get a 'sync' event on the watch stream), but those
are handled gracefully already.

[1] https://github.com/kubernetes/community/blob/master/contributors/devel/api-conventions.md#concurrency-control-and-consistency