Retry (supposedly) transient network errors infinitely
Instead of leaving a specific queue in the 'rejected' state when transient network errors such
as "host unreachable" or "connection refused" occur we're now resolving the promise with another
promise that will retry the same operation. This still ensures that we keep the order of operations
consistent, but and may still make us never do anything, but gives us a better chance to actually
resolving the pending work items properly.

NB: The other option here would have been to actually abort the whole process, and hope that the
    issue would be gone by the next time kubernetes restarted it.