Rewrite the documentation to be easier to read
The aws-resource-service is intended for people that understand the AWS resources involved, and
how a "operator" in general works. The documentation we _need_ to have is how one should create
the resources, under the assumption that the AWS documentation (CloudFormation or SDK) is available
for consultation.

There is no point in trying to list all possible properties, as in most cases we don't care at all,
and just pass these to the underlying AWS SDK for processing.