Make capitalizeFieldNames handle arrays correctly
We used to be able to handle arrays that were values of a property of an object, but still
fail on arrays provided as "root" of the capitalization. Fixing this is reasonably easy by moving
the check on the "array-ness" of the input up to the point where we determine the output type:
Regardless of an array or object we can reduce over the keys (for arrays these will be the indices),
but depending on the type we need to start with the correct type of object for that reduction
operation.