Restore the "declarative" aspect for IAM tags
The implemented logic would fail to remove tags when the `spec` field would not specify them. This breaks
one of the core ideas of this service:
> The `spec` leads, and the service adjusts the AWS resource state to what the `spec` describes

This commit makes sure to translate a missing `tags` property into an empty array, and then always
calls `tagRole` to update the role accordingly.