Simplify the `TranslateAttributesResult` and related code for the S3 resource
This reduces the complexity in kubernetes-to-aws by only translating our/CloudFormations structure
into the needed type for the S3 request, but not bothering with particulars on how to invoke the S3
APIs itself (i.e. the bucket name and content-md5 parts that are required for most requests).

This separates the responsibilities, and simplifies a lot of code later on in the aws module: Because
there are no "BucketName" properties in the configuration objects that can be different from the bucketName
parameters in the function call we simply don't need to check these conditions.