Ensure that changes in the Kubernetes resource lead the state of the AWS resource
The aws-resource-service is responsible for translating the current kubernetes "spec" into
instructions that bring the AWS resource into that state. This means that in the update-logic
we need to correctly consider the case where a spec property was removed.

This commit fixes these problems for:
* logging: we now correctly suspend logging when the logging configuration dropped
* versioning: depending on how the current bucket versioning configuration looks like
  versioning will be disabled when it is not provided in the spec.
* lifecycle rules: all rules will be deleted when the update doesn't specify any
* tagging: all tags will be removed when the update doesn't specify any tags

This commit also adds some unit tests for the touched areas. These tests should work in `npm test`, but
do have a "prototype pollution problem" that will make the S3Client tests fail when the utils test has
executed in the same Node VM (i.e. with the same require cache).